requirements.txt
discord.py
Pillow
python-dotenv
requests
aiosqlite